refactor: 拆分 defget 纯工具并缩小导入面
