+++ src/utils/jjc_text.py
@@ -1,0 +1,24 @@
+from __future__ import annotations
+
+
+def jjcdaxiaoxie(timestamp: int) -> str | None:
+    mapping = {
+        0: "零",
+        1: "一",
+        2: "二",
+        3: "三",
+        4: "四",
+        5: "五",
+        6: "六",
+        7: "七",
+        8: "八",
+        9: "九",
+        10: "十",
+        11: "十一",
+        12: "十二",
+        13: "十三",
+        14: "十四",
+        15: "十五",
+    }
+    return mapping.get(timestamp)
+
